// payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
// ...
export interface FeaturedPropertiesBlock {
  communityOne: number | Community;
  communityTwo: number | Community;
  communityThree: number | Community;
  id?: string | null;
  blockName?: string | null;
  blockType: 'featuredProperties';
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "communities".
 */
export interface Community {
  id: number;
  name: string;
  address: {
    addressLine1: string;
    addressLine2?: string | null;
    city: string;
    state: string;
    zip: string;
  };
  description: string;
  websiteUrl: string;
  fund: string;
  yearAcquired: string;
  units: string;
  sold?: boolean | null;
  features: {
    featureOne: string;
    featureTwo: string;
    featureThree: string;
  };
  featuredImage: number | Media;
  images: (number | Media)[];
  publishedAt?: string | null;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}